requirements.txt
amqp==5.3.1
asgiref==3.10.0
attrs==25.4.0
audioop-lts==0.2.2
audioread==3.1.0
billiard==4.2.4
boto3==1.40.69
botocore==1.40.69
celery==5.6.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
coverage==7.11.3
cron_descriptor==2.0.6
decorator==5.2.1
deprecation==2.1.0
dj-database-url==1.2.0
Django==5.2.8
django-celery-beat==2.8.1
django-cors-headers==4.3.0
django-environ==0.12.0
django-extensions==4.1
django-filter==25.2
django-redis==6.0.0
django-storages==1.14.6
django-timezone-field==7.2.1
django_celery_results==2.6.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
drf-yasg==1.21.11
eyeD3==0.9.9
factory_boy==3.3.3
Faker==38.2.0
filetype==1.2.0
gevent==24.10.3
greenlet==3.1.1
gunicorn==21.2.0
idna==3.11
inflection==0.5.1
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.3
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.6.2
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.46.0
MarkupSafe==3.0.3
msgpack==1.1.2
mutagen==1.47.0
numba==0.63.1
numpy==2.3.5
packaging==25.0
pillow==12.0.0
platformdirs==4.5.1
pooch==1.8.2
prompt_toolkit==3.0.52
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg2==2.9.11
psycopg2-binary==2.9.11
pycparser==3.0
pydub==0.25.1
PyJWT==2.10.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==2.0.7
pytz==2025.2
PyYAML==6.0.3
redis==7.1.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.28.0
s3transfer==0.14.0
scikit-learn==1.8.0
scipy==1.17.0
setuptools==80.9.0
six==1.17.0
soundfile==0.13.1
soxr==1.0.0
sqlparse==0.5.3
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
swagger_ui_bundle==1.1.0
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.3.3
wheel==0.45.1
whitenoise==6.5.0
zope.event==6.1
zope.interface==8.2
qrcode[pil]==7.4.2 
